+++ cloud/azure/common/deploy/cron.go
